introduced runWithProgress template function
