+++ cmd/restic/cmd_backup.go
@@ -19,7 +19,6 @@ import (
 	"github.com/restic/restic/internal/restic"
 	"github.com/restic/restic/internal/textfile"
 	"github.com/restic/restic/internal/ui"
-	"github.com/restic/restic/internal/ui/termstatus"
 	"github.com/spf13/cobra"
 	tomb "gopkg.in/tomb.v2"
 )
@@ -51,26 +50,7 @@ given as the arguments.
 			}
 		}
 
-		// XXX this is EXACT copy from cmd_restore, and all other commands will need this too
-		var t tomb.Tomb
-		term := termstatus.New(globalOptions.stdout, globalOptions.stderr, globalOptions.Quiet)
-		t.Go(func() error { term.Run(t.Context(globalOptions.ctx)); return nil })
-
-		prevStdout, prevStderr := globalOptions.stdout, globalOptions.stderr
-		defer func() {
-			globalOptions.stdout, globalOptions.stderr = prevStdout, prevStderr
-		}()
-		pm := ui.NewTermstatusProgressUI(term, globalOptions.verbosity)
-		defer pm.Finish()
-		globalOptions.stdout, globalOptions.stderr = pm.Stdout(), pm.Stderr()
-		t.Go(func() error { return pm.Run(t.Context(globalOptions.ctx)) })
-
-		err := runBackup(backupOptions, globalOptions, pm, args)
-		if err != nil {
-			return err
-		}
-		t.Kill(nil)
-		return t.Wait()
+		return runWithProgress(func(pm ui.ProgressUI) error { return runBackup(backupOptions, globalOptions, pm, args) })
 	},
 }
