+++ cmd/restic/cmd_restore.go
@@ -7,9 +7,6 @@ import (
 	"github.com/restic/restic/internal/restic"
 	"github.com/restic/restic/internal/restorer"
 	"github.com/restic/restic/internal/ui"
-	"github.com/restic/restic/internal/ui/termstatus"
-	tomb "gopkg.in/tomb.v2"
-
 	"github.com/spf13/cobra"
 )
 
@@ -25,20 +22,7 @@ repository.
 `,
 	DisableAutoGenTag: true,
 	RunE: func(cmd *cobra.Command, args []string) error {
-		var t tomb.Tomb
-		term := termstatus.New(globalOptions.stdout, globalOptions.stderr, globalOptions.Quiet)
-		t.Go(func() error { term.Run(t.Context(globalOptions.ctx)); return nil })
-
-		prevStdout, prevStderr := globalOptions.stdout, globalOptions.stderr
-		defer func() {
-			globalOptions.stdout, globalOptions.stderr = prevStdout, prevStderr
-		}()
-		pm := ui.NewTermstatusProgressUI(term, globalOptions.verbosity)
-		defer pm.Finish()
-		globalOptions.stdout, globalOptions.stderr = pm.Stdout(), pm.Stderr()
-		t.Go(func() error { return pm.Run(t.Context(globalOptions.ctx)) })
-
-		return runRestore(restoreOptions, globalOptions, pm, args)
+		return runWithProgress(func(pm ui.ProgressUI) error { return runRestore(restoreOptions, globalOptions, pm, args) })
 	},
 }
